requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
